Extract sync delta helpers
