+++ service-worker.js
@@ -139,6 +139,7 @@ const APP_SHELL = [
   '/js/hardware.js',
   '/js/sync.js',
   '/js/sync-apply.js',
+  '/js/sync-delta.js',
   '/js/sync-payload.js',
   '/js/sync-relay-health.js',
   '/js/sync-state.js',
